// Web/RMA-Web-Shell/projects/claimcare/src/app/claim-manager/shared/enums/disease-type-enum.ts
export enum DiseaseTypeEnum {
    Unknown = 0,
    NIHL = 1,
    Dermatitis = 2,
    TuberculosisHealthWorkersOnly = 3,
    OccupationalAsthma = 4,
    Silicosis = 5,
    Exposure = 6,
    WorkRelatedUpperLimbDisorder = 7,
    PulmonaryTuberculosis = 8,
    Pneumoconiosis = 9,
    ObstructiveAirwaysDisease = 10,
    OtherMBODCompensableDisease = 11,
    PulmonaryTuberculosisPlusPneumoconiosis = 12,
    PulmonaryTuberculosisPlusOAD = 13,
    PulmonaryTuberculosisPlusPneumoconiosisPlusOAD = 14,
    PneumoconiosisPlusObstructiveAirwaysDisease = 15,
    HeatExhaustionHeatStroke = 16,
    CardioPulmonaryTuberculosis = 17,
    TuberculousPleurisy = 18,
    TuberculosisOfTheHeart = 19,
    AsbestosisWithPleuralAsbestosis = 20,
    AsbestosisInterstitialAndPleural = 21,
    MesotheliumOfPleura = 22,
    MesotheliumOfPeritoneum = 23,
    MesotheliumOfPericardium = 24,
    AsbestosRelatedLungCancer = 25,
    CoalWorkerPneumoconiosis = 26,
    PlatinumSaltSensitivity = 27,
    Virus = 28,
    Malaria = 29,
    PTSD = 30,
    PulmonaryTB = 31,
		COVID19 = 32,
  }